🔥 llm 로직 gpt앱 -> task앱으로 변경 (#24)
