+++ project/scandy/urls.py
@@ -26,7 +26,7 @@ urlpatterns = [
     path('session/', include('apps.session_tokens.urls')),
     path('google_calendar/', include('apps.google_calendar.urls')),
     path('task/', include('apps.tasks.urls')),
-    path('gpt/', include('apps.gpt.urls')),
+    # path('gpt/', include('apps.gpt.urls')),
     
     path('api/schema/', SpectacularJSONAPIView.as_view(), name='schema-json'),
     path('api/redoc/', SpectacularRedocView.as_view(url_name='schema-json'), name='redoc'),
